requirements.txt
python-telegram-bot==5.0.0
configparser==3.5.0
